move utils folder to src folder
